requirements.txt
black==19.10b0
Django==3.0.7
django-cors-headers==3.3.0
django-oauth-toolkit==1.3.2
djangorestframework==3.11.0
psycopg2==2.8.4
pytest==5.3.1
pytest-django==3.8.0
pytz==2019.3
requests==2.22.0
uuid==1.30
